+++ packages/presentation/src/utils.ts
@@ -524,3 +524,18 @@ export function getFiltredKeys (
 export function isCollectionAttr (hierarchy: Hierarchy, key: KeyedAttribute): boolean {
   return hierarchy.isDerived(key.attr.type._class, core.class.Collection)
 }
+
+/**
+ * @public
+ */
+export function decodeTokenPayload (token: string): any {
+  const parts = token.split('.')
+
+  const payload = parts[1]
+
+  const decodedPayload = atob(payload)
+
+  const parsedPayload = JSON.parse(decodedPayload)
+
+  return parsedPayload
+}
