+++ plugins/setting-resources/src/components/WorkspaceSettings.svelte
@@ -15,7 +15,7 @@
 <script lang="ts">
   import { PersonAccount } from '@hcengineering/contact'
   import { AccountRole, getCurrentAccount, roleOrder } from '@hcengineering/core'
-  import { createQuery } from '@hcengineering/presentation'
+  import presentation, { createQuery, decodeTokenPayload } from '@hcengineering/presentation'
   import setting, { SettingsCategory } from '@hcengineering/setting'
   import {
     Component,
@@ -27,6 +27,7 @@
   } from '@hcengineering/ui'
   import { onDestroy } from 'svelte'
   import { clearSettingsStore } from '../store'
+  import { getMetadata } from '@hcengineering/platform'
 
   export let kind: 'navigation' | 'content' | undefined
   export let categoryName: string
@@ -38,12 +39,17 @@
   let categories: SettingsCategory[] = []
   const account = getCurrentAccount() as PersonAccount
 
+  const admin = decodeTokenPayload(getMetadata(presentation.metadata.Token) ?? '').admin === 'true'
+
   const settingsQuery = createQuery()
   settingsQuery.query(
     setting.class.WorkspaceSettingCategory,
     {},
     (res) => {
       categories = roleOrder[account.role] > roleOrder[AccountRole.User] ? res : res.filter((p) => !p.secured)
+      if (!admin) {
+        categories = categories.filter((p) => !p.adminOnly)
+      }
       category = findCategory(categoryId)
     },
     { sort: { order: 1 } }
